fix(model): Keep the order of 1:n lists by reversing the id traversal
Before this commits the list of related entries was collected by traversing the included entries which does not care about the order of the ids coming from the contentful API, in fact using this approach the related entries were sorted by id. This commit changes the approach to iterate over the entryIDs of the relation field that has the 1:n related entries. Quite an easy fix with a big feature, since content editors will expect lists to be in the order arranged in contentful when viewing on the using website/app.